back/requirements.txt
flask
flask-cors
pymongo
gunicorn
google-cloud-secret-manager